refactor: rename form handling function and update import paths for consistency
